chatbot: move Image to a shared types::media module
Image is a generic media type, not conversation-specific. Lift it out of
types::conversation into its own types::media module (also a natural home
for Audio later, since the mtmd context advertises vision and audio).
