+++ chatbot/src/types.rs
@@ -1,1 +1,2 @@
 pub mod conversation;
+pub mod media;
