+++ chatbot/src/types/media.rs
@@ -1,0 +1,17 @@
+use serde::{Deserialize, Serialize};
+use std::sync::Arc;
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct Image {
+    pub bytes: Arc<Vec<u8>>,
+    pub mime: String,
+}
+
+impl std::fmt::Debug for Image {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Image")
+            .field("mime", &self.mime)
+            .field("bytes", &format_args!("{} bytes", self.bytes.len()))
+            .finish()
+    }
+}
